% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scico-package.R
\docType{package}
\name{scico-package}
\alias{scico-package}
\alias{_PACKAGE}
\title{scico: Colour Palettes Based on the Scientific Colour-Maps}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Colour choice in information visualisation is important in order to avoid being mislead by inherent bias in the used colour palette. The 'scico' package provides access to the perceptually uniform and colour-blindness friendly palettes developed by Fabio Crameri and released under the "Scientific Colour-Maps" moniker. The package contains 24 different palettes and includes both diverging and sequential types.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/thomasp85/scico}
  \item Report bugs at \url{https://github.com/thomasp85/scico/issues}
}

}
\author{
\strong{Maintainer}: Thomas Lin Pedersen \email{thomasp85@gmail.com} (\href{https://orcid.org/0000-0002-5147-4711}{ORCID})

Authors:
\itemize{
  \item Fabio Crameri
}

}
