% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{scico_data}
\alias{scico_data}
\alias{scico_palette_names}
\alias{scico_palette_data}
\title{Access raw palette data}
\usage{
scico_palette_names()

scico_palette_data(palette)
}
\arguments{
\item{palette}{The name of the palette to fetch data for}
}
\description{
These functions gives access to the data underlying the palettes
}
\keyword{internal}
